+++ sdk/keyvault/Azure.Security.KeyVault.Administration/src/KeyVaultAdministrationClientOptions.cs
@@ -18,7 +18,7 @@ namespace Azure.Security.KeyVault.Administration
         /// For more information, see
         /// <see href="https://docs.microsoft.com/rest/api/keyvault/key-vault-versions">Key Vault versions</see>.
         /// </summary>
-        internal const ServiceVersion LatestVersion = ServiceVersion.V7_5;
+        internal const ServiceVersion LatestVersion = ServiceVersion.V7_6_Preview_1;
 
         /// <summary>
         /// The versions of Azure Key Vault supported by this client
@@ -46,6 +46,11 @@ namespace Azure.Security.KeyVault.Administration
             /// The Key Vault API version 7.5.
             /// </summary>
             V7_5 = 4,
+
+            /// <summary>
+            ///  The Key Vault API version 7.6-preview.1.
+            /// </summary>
+            V7_6_Preview_1 = 5,
 #pragma warning restore CA1707 // Identifiers should not contain underscores
         }
 
@@ -83,6 +88,7 @@ namespace Azure.Security.KeyVault.Administration
                 ServiceVersion.V7_3 => "7.3",
                 ServiceVersion.V7_4 => "7.4",
                 ServiceVersion.V7_5 => "7.5",
+                ServiceVersion.V7_6_Preview_1 => "7.6-preview.1",
                 _ => throw new ArgumentException(Version.ToString()),
             };
         }