+++ sdk/keyvault/Azure.Security.KeyVault.Administration/src/OperationHeadersExtensions.cs
@@ -17,6 +17,24 @@ namespace Azure.Security.KeyVault.Administration
         }
 
         /// <summary>
+        /// Extracts the operation JobId from the <see cref="AzureSecurityKeyVaultAdministrationPreFullBackupHeaders"/> AzureAsyncOperation.
+        /// </summary>
+        /// <returns>The operation JobId.</returns>
+        public static string JobId(this AzureSecurityKeyVaultAdministrationPreFullBackupHeaders header)
+        {
+            return GetJobIdFromAzureAsyncOperation(header.AzureAsyncOperation);
+        }
+
+        /// <summary>
+        /// Extracts the operation JobId from the <see cref="AzureSecurityKeyVaultAdministrationPreFullRestoreOperationHeaders"/> AzureAsyncOperation.
+        /// </summary>
+        /// <returns>The operation JobId.</returns>
+        public static string JobId(this AzureSecurityKeyVaultAdministrationPreFullRestoreOperationHeaders header)
+        {
+            return GetJobIdFromAzureAsyncOperation(header.AzureAsyncOperation);
+        }
+
+        /// <summary>
         /// Extracts the operation JobId from the <see cref="AzureSecurityKeyVaultAdministrationFullRestoreOperationHeaders" /> AzureAsyncOperation.
         /// </summary>
         /// <returns>The operation JobId.</returns>